pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-contract-review"
version = "0.1.0"
description = "LangGraph Studio contract review workflow with human approval and audit logging."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "langgraph>=1.0.0",
  "pydantic>=2.8.0",
  "pypdf>=5.0.0",
  "python-docx>=1.1.2",
  "pyyaml>=6.0.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["contract_review_langgraph*"]

